+++ src/main/java/com/PBA/authservice/persistance/repository/PendingUserDaoImpl.java
@@ -3,6 +3,7 @@ package com.pba.authservice.persistance.repository;
 import com.pba.authservice.persistance.model.PendingUser;
 import com.pba.authservice.persistance.repository.mappers.PendingUserRowMapper;
 import com.pba.authservice.persistance.repository.sql.PendingUserSqlProvider;
+import com.pba.authservice.persistance.repository.sql.PendingUserSqlProviderImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
