+++ src/main/java/com/PBA/authservice/service/ActiveUserServiceImpl.java
@@ -1,16 +1,17 @@
 package com.pba.authservice.service;
 
 import com.pba.authservice.persistance.model.ActiveUser;
+import com.pba.authservice.persistance.repository.ActiveUserDao;
 import com.pba.authservice.persistance.repository.ActiveUserDaoImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class ActiveUserServiceImpl implements ActiveUserService {
-    private final ActiveUserDaoImpl activeUserDao;
+    private final ActiveUserDao activeUserDao;
 
     @Autowired
-    public ActiveUserServiceImpl(ActiveUserDaoImpl activeUserDao) {
+    public ActiveUserServiceImpl(ActiveUserDao activeUserDao) {
         this.activeUserDao = activeUserDao;
     }
