+++ src/test/java/com/pba/authservice/unit/ActiveUserDaoUnitTest.java
@@ -3,10 +3,10 @@ package com.pba.authservice.unit;
 import com.pba.authservice.exceptions.AuthDaoException;
 import com.pba.authservice.mockgenerators.ActiveUserMockGenerator;
 import com.pba.authservice.persistance.model.ActiveUser;
-import com.pba.authservice.persistance.repository.ActiveUserDao;
 import com.pba.authservice.persistance.repository.ActiveUserDaoImpl;
 import com.pba.authservice.persistance.repository.mappers.ActiveUserRowMapper;
 import com.pba.authservice.persistance.repository.sql.ActiveUserSqlProvider;
+import com.pba.authservice.persistance.repository.sql.ActiveUserSqlProviderImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
