+++ src/test/java/com/pba/authservice/unit/PendingUserDaoUnitTest.java
@@ -3,10 +3,10 @@ package com.pba.authservice.unit;
 import com.pba.authservice.exceptions.AuthDaoException;
 import com.pba.authservice.mockgenerators.PendingUserMockGenerator;
 import com.pba.authservice.persistance.model.PendingUser;
-import com.pba.authservice.persistance.repository.PendingUserDao;
 import com.pba.authservice.persistance.repository.PendingUserDaoImpl;
 import com.pba.authservice.persistance.repository.mappers.PendingUserRowMapper;
 import com.pba.authservice.persistance.repository.sql.PendingUserSqlProvider;
+import com.pba.authservice.persistance.repository.sql.PendingUserSqlProviderImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
