+++ src/test/resources/application.yml
@@ -1,5 +1,5 @@
 spring:
   datasource:
-    url: jdbc:postgresql://localhost:5432/testdb
+    url: jdbc:postgresql://localhost:5432/authtest
     username: postgres
     password: larlarlar
